MDSynthesis now uses datreant objects under the hood!
Development of datreant can take place without worrying about
MD-specific concernts, while MDS builds on those components for MD uses.
The MDS tests all pass once more, and the package is now a lot smaller.

MDS development will trail datreant development, with releases of MDS
corresponding to already-released versions of datreant. This wil make it
easier for MDS to follow what is essentially a moving target without
having to follow it commit-by-commit. There may not necessarily be
a release of MDS for every release of datreant, however.

Something to consider: at present Sims are kind of monkey-patched
into datreant functions through mdsynthesis.__init__. The current way
this is done may not be terribly robust, so it's worth investigating
limitations and alternatives.