After much labor, we now have everything working with JSON files.
We even included the resnum functionality, since we don't really lose
much by including it for now, and we keep compatibility with a lot of
existing code that actually uses this. Eventually, once MDAnalysis'
Topology refactor is in full effect, we can defer persistence of things
like resnums to that object.