Moved universe definition methods to own limb, changed selections to atomselections
We have to be careful that the namespace of a Sim doesn't become a
liability in the future depending on what goes on in datreant. This is
one issue with the approach datreant takes with respect to limb-based
functionality: namespace conflicts are very possible, and I don't have a
good policy in mind for how to resolve this long-term. At the moment
there's no issue because we don't have much going on, but we have to
mitigate the issue now so it's not hell later.

Moving universe definition properties/methods into their own limb limits
the possibility of namespace conflicts, since it's one name (udef) vs.
meany introduced into the Treant's top-level namespace. It also means
that later we can treat the universe definition as an object, and so we
can make a setter a la Tags for duplicating a universe definition
between Sims.

Also, changed selections to atomselections in anticipation of the
future. "selections" is too generic a name to use in an
application-specific package, as this could get used by a datreant
package later.