Sim.atomselections no longer triggers Universe build on call
It can be expensive to build some Universes, so we want to be able to
store selections without having it built. This removes a relic from the
old "multiple Universes" scheme, in which the default Universe was
loaded to make it the "active" Universe for any selections work if no
Universe was already active.