No longer need a separate mds.Bundle.
Minor fixes in datreant.core.Bundle allow for Sims to be handled
gracefully, so long as they are passed in as objects and not paths.
I think this is acceptable because first and foremost, a Sim is a
Treant, so passing in paths only will give you a Bundle of base Treants.

Slicing and filtering on Bundles will always yield Bundles with the
class of the members preserved, since these are held onto in the
Bundle's cache and passed around.